feat(service_unraid): split the screen into tabs and show per-core load
One scroll was already long with the array, the load graphs and the
containers on it, and virtual machines would have made it worse. Four
tabs now, matching the other services with this much to show.

Its navigation bar reserves the system inset beneath it. Material's
NavigationBar wraps itself in a SafeArea and then sizes to its height, so
a bare 80 leaves the bar eating that inset out of itself: measured on a
three-button device the inset is 48, which left 32 for icons and labels
and squeezed them together. The bar lives in core_ui as a widget with a
test that fails if the inset stops being reserved, rather than as another
copy of the same AnimatedContainer.

Per core CPU sits under the average, because the average is what hides
the case worth seeing: one core pinned while the others idle reads as a
quiet machine, and on a sixteen thread box it averages out to six
percent. Bars side by side make that shape obvious.

Virtual machines get a tab of their own, read only for now. The API has
start, stop and the rest, but nothing available to test against runs a
VM, and buttons nobody has watched work are worse than a list. Unraid
ships with virtualisation off and refuses the query outright rather than
answering with nothing, so that refusal is reported as a state instead of
raised, and asked for on its own: an errors array anywhere in a response
fails the whole request, which would have taken the array and the
containers down with it. A stopped Docker service does the same thing and
is named the same way.
