+++ packages/core_ui/pubspec.yaml
@@ -23,3 +23,5 @@ dependencies:
 
 dev_dependencies:
   build_runner: ^2.4.13
+  flutter_test:
+    sdk: flutter
