+++ services/service_unraid/lib/service_unraid.dart
@@ -2,6 +2,10 @@
 library;
 
 export 'src/models/unraid_models.dart';
+export 'src/tabs/unraid_array_tab.dart';
+export 'src/tabs/unraid_docker_tab.dart';
+export 'src/tabs/unraid_system_tab.dart';
+export 'src/tabs/unraid_vms_tab.dart';
 export 'src/unraid_client.dart';
 export 'src/unraid_home.dart';
 export 'src/unraid_providers.dart';
